scripts/requirements.txt
SQLAlchemy==2.0.23
PyMySQL==1.1.0
python-dotenv==1.0.0

